requirements.txt
discord.py
black
humanfriendly
pytz
tortoise-orm[asyncpg]